third_party/WebKit/LayoutTests/http/tests/devtools/debugger/fetch-breakpoints-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests fetch() breakpoints.


Running: testFetchBreakpoint
Script execution paused.
Call stack:
    0) sendRequest (fetch-breakpoints.js:11)
    1)  (:1)
Script execution resumed.

Running: testPauseOnAnyFetch
Script execution paused.
Script execution resumed.
Script execution paused.
Script execution resumed.

Running: testDisableBreakpoint
Script execution paused.
Script execution resumed.
